+++ apps/backend/src/models.rs
@@ -7,8 +7,7 @@ use specta::Type;
 
 use crate::{
     entities::exercise::Model as ExerciseModel,
-    migrator::{MetadataLot, MetadataSource, SeenState},
-    miscellaneous::{MediaSpecifics, MetadataCreator, MetadataImage},
+    migrator::{MetadataImageLot, MetadataLot, MetadataSource, SeenState},
 };
 
 #[derive(Debug, Serialize, Deserialize, Clone, SimpleObject, InputObject)]
@@ -569,6 +568,84 @@ pub mod media {
         pub reviews: Vec<ImportOrExportItemRating>,
         /// The collections to add this media to.
         pub collections: Vec<String>,
+    }
+
+    #[derive(Debug, Serialize, Deserialize, Clone, FromJsonQueryResult, Eq, PartialEq, Default)]
+    #[serde(tag = "t", content = "d")]
+    pub enum MediaSpecifics {
+        AudioBook(AudioBookSpecifics),
+        Book(BookSpecifics),
+        Movie(MovieSpecifics),
+        Podcast(PodcastSpecifics),
+        Show(ShowSpecifics),
+        VideoGame(VideoGameSpecifics),
+        Anime(AnimeSpecifics),
+        Manga(MangaSpecifics),
+        #[default]
+        Unknown,
+    }
+
+    #[derive(Debug, Serialize, Deserialize, Clone, Eq, PartialEq, Hash)]
+    pub enum MetadataImageUrl {
+        S3(String),
+        Url(String),
+    }
+
+    impl Default for MetadataImageUrl {
+        fn default() -> Self {
+            Self::Url("".to_owned())
+        }
+    }
+
+    #[derive(
+        Clone, Debug, PartialEq, FromJsonQueryResult, Eq, Serialize, Deserialize, Default, Hash,
+    )]
+    pub struct MetadataImage {
+        pub url: MetadataImageUrl,
+        pub lot: MetadataImageLot,
+    }
+
+    #[derive(Clone, Debug, PartialEq, FromJsonQueryResult, Eq, Serialize, Deserialize, Default)]
+    pub struct MetadataImages(pub Vec<MetadataImage>);
+
+    #[derive(
+        Clone,
+        Debug,
+        PartialEq,
+        FromJsonQueryResult,
+        Eq,
+        Serialize,
+        Deserialize,
+        SimpleObject,
+        Default,
+        Hash,
+    )]
+    pub struct MetadataCreator {
+        pub name: String,
+        pub role: String,
+        pub image_urls: Vec<String>,
+    }
+
+    #[derive(
+        Clone, Debug, PartialEq, FromJsonQueryResult, Eq, Serialize, Deserialize, Default, Hash,
+    )]
+    pub struct MetadataCreators(pub Vec<MetadataCreator>);
+
+    #[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Clone, SimpleObject)]
+    pub struct SeenShowExtraInformation {
+        pub season: i32,
+        pub episode: i32,
+    }
+
+    #[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Clone, SimpleObject)]
+    pub struct SeenPodcastExtraInformation {
+        pub episode: i32,
+    }
+
+    #[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Clone, FromJsonQueryResult)]
+    pub enum SeenOrReviewExtraInformation {
+        Show(SeenShowExtraInformation),
+        Podcast(SeenPodcastExtraInformation),
     }
 }