+++ apps/backend/src/providers/google_books.rs
@@ -8,9 +8,11 @@ use surf::{http::headers::ACCEPT, Client};
 use crate::{
     config::GoogleBooksConfig,
     migrator::{MetadataImageLot, MetadataLot, MetadataSource},
-    miscellaneous::{MediaSpecifics, MetadataCreator, MetadataImage, MetadataImageUrl},
     models::{
-        media::{BookSpecifics, MediaDetails, MediaSearchItem},
+        media::{
+            BookSpecifics, MediaDetails, MediaSearchItem, MediaSpecifics, MetadataCreator,
+            MetadataImage, MetadataImageUrl,
+        },
         SearchResults,
     },
     traits::{MediaProvider, MediaProviderLanguages},