+++ apps/frontend/src/lib/constants.ts
@@ -3,18 +3,27 @@ export const ROUTES = {
 		login: "/auth/login",
 		register: "/auth/register",
 	},
-	collections: {
-		list: "/collections/list",
-		details: "/collections",
-	},
-	dashboard: "/",
-	list: "/list",
-	media: {
-		commit: "/media/commit",
-		create: "/media/create",
-		details: "/media",
-		postReview: "/media/post-review",
-		updateProgress: "/media/update-progress",
-	},
 	settings: "/settings",
+	imports: {
+		new: "/imports",
+		reports: "/imports/reports",
+	},
+	media: {
+		collections: {
+			list: "/collections/list",
+			details: "/collections",
+		},
+		dashboard: "/",
+		list: "/list",
+		individualMedia: {
+			commit: "/media/commit",
+			create: "/media/create",
+			details: "/media",
+			postReview: "/media/post-review",
+			updateProgress: "/media/update-progress",
+		},
+	},
+	fitness: {
+		home: "/fitness",
+	},
 } as const;